Update schema and tests
