+++ src/backend/tests/services/test_film_service.py
@@ -65,8 +65,6 @@ def test_get(app, add_test_film):
             "weeks": [
                 {
                     "date": "2022-01-20",
-                    "film": "Nope",
-                    "film_slug": "nope",
                     "id": 1,
                     "week_gross": 1000,
                     "weekend_gross": 500,
