+++ src/backend/ukbo/dto/FilmSchema.py
@@ -4,7 +4,17 @@ from ukbo.extensions import ma
 
 from .CountrySchema import CountrySchema
 from .DistributorSchema import DistributorSchema
-from .FilmWeekSchema import FilmWeekSchema
+
+
+class FilmWeekSchemaStrict(ma.SQLAlchemyAutoSchema):
+    """
+    FilmWeek schema for serialisation, not including Film object.
+    """
+
+    class Meta:
+        model = models.Film_Week
+
+    date = ma.Function(lambda obj: obj.date.strftime("%Y-%m-%d"))
 
 
 class FilmSchema(ma.SQLAlchemyAutoSchema):
@@ -18,7 +28,7 @@ class FilmSchema(ma.SQLAlchemyAutoSchema):
 
     distributors = ma.Nested(DistributorSchema, many=True)
     countries = ma.Nested(CountrySchema, many=True)
-    weeks = ma.Nested(FilmWeekSchema, many=True)
+    weeks = ma.Nested(FilmWeekSchemaStrict, many=True)
     gross = ma.Function(lambda obj: obj.gross)
 
 
@@ -47,3 +57,32 @@ class FilmSchemaValues(ma.SQLAlchemyAutoSchema):
 
     label = fields.String(attribute="name")
     value = fields.String(attribute="id")
+
+
+class FilmWeekSchema(ma.SQLAlchemyAutoSchema):
+    """
+    FilmWeek schema for serialisation.
+    """
+
+    class Meta:
+        model = models.Film_Week
+
+    date = ma.Function(lambda obj: obj.date.strftime("%Y-%m-%d"))
+    film = ma.Nested(FilmSchemaStrict)
+
+
+class FilmWeekSchemaArchive(ma.SQLAlchemyAutoSchema):
+    """
+    FilmWeek Archive schema for serialisation.
+    Strictly for the archive endpoint.
+    """
+
+    class Meta:
+        model = models.Film_Week
+
+    date = ma.Function(lambda obj: obj.date.strftime("%Y%m%d"))
+    film = ma.Function(lambda obj: obj.film.name)
+    distributor = ma.Function(
+        lambda obj: [x.name for x in obj.film.distributors]
+    )
+    country = ma.Function(lambda obj: [x.name for x in obj.film.countries])
