+++ src/backend/ukbo/dto/__init__.py
@@ -9,6 +9,11 @@ DTOs are used to serialize the data that is sent to the API.
 from .CountrySchema import CountrySchema
 from .DistributorSchema import DistributorSchema
 from .EventSchema import EventSchema
-from .FilmSchema import FilmSchema, FilmSchemaStrict, FilmSchemaValues
-from .FilmWeekSchema import FilmWeekSchema, FilmWeekSchemaArchive
+from .FilmSchema import (
+    FilmSchema,
+    FilmSchemaStrict,
+    FilmSchemaValues,
+    FilmWeekSchema,
+    FilmWeekSchemaArchive,
+)
 from .WeekSchema import WeekSchema
